requirements.txt
streamlit
openai>=1.0.0
pillow
numpy
python-dotenv    
ollama
requests